Fixed avatar rendering condition
